Factor out hunk diff body
Create a stand-alone component called `HunkDiffBody` in order to be able to reuse the syntax highlighting of the diff patches quoted on messages
